Rimossi check null, continuo autenticazione, aggiunte librerie

+++ src/backend/main/java/com/retrobased/market/services/ServiceIndirizzoCliente.java
@@ -3,9 +3,11 @@ package com.retrobased.market.services;
 import com.retrobased.market.entities.IndirizzoCliente;
 import com.retrobased.market.repositories.RepositoryCliente;
 import com.retrobased.market.repositories.RepositoryIndirizzoCliente;
+import com.retrobased.market.support.exceptions.AddressNotExist;
 import com.retrobased.market.support.exceptions.ClientTokenMismatch;
 import com.retrobased.market.support.exceptions.ClientNotExist;
 import com.retrobased.market.support.exceptions.ValueCannotBeEmpty;
+import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,10 +25,8 @@ public class ServiceIndirizzoCliente {
     }
 
     @Transactional(propagation = Propagation.REQUIRED)
-    public IndirizzoCliente addAddress(IndirizzoCliente address) throws ClientNotExist, ValueCannotBeEmpty {
-        if (address == null
-                || address.getId() == null
-                || address.getCittà() == null
+    public IndirizzoCliente addAddress(@NonNull Integer idCliente, @NonNull IndirizzoCliente address) throws ClientNotExist, ValueCannotBeEmpty {
+        if (address.getCittà() == null
                 || address.getPaese() == null
                 || address.getLineaIndirizzo1() == null
                 || address.getCodicePostale() == null
@@ -37,15 +37,16 @@ public class ServiceIndirizzoCliente {
         if (userRepo.NotExistById(address.getIdCliente().getId()))
             throw new ClientNotExist();
 
+        address.setIdCliente(userRepo.getReferenceById(idCliente));
+
         return addressRepo.save(address);
 
     }
 
     // TODO CAMBIARE IDCLIENTE CON TOKEN
     @Transactional(propagation = Propagation.REQUIRED)
-    public void removeAddress(Integer idClient, IndirizzoCliente address) throws ClientNotExist, ValueCannotBeEmpty, ClientTokenMismatch {
-        if (address == null
-                || address.getId() == null
+    public void removeAddress(@NonNull Integer idClient, @NonNull IndirizzoCliente address) throws ClientNotExist, ValueCannotBeEmpty, ClientTokenMismatch, AddressNotExist {
+        if (address.getId() == null
                 || address.getIdCliente() == null
         )
             throw new ValueCannotBeEmpty();
@@ -56,6 +57,8 @@ public class ServiceIndirizzoCliente {
         if (userRepo.NotExistById(address.getIdCliente().getId()))
             throw new ClientNotExist();
 
+        if(!addressRepo.existsIndirizzoClienteByIdAndIdCliente(address.getId(), idClient))
+           throw new AddressNotExist();
 
         userRepo.deleteById(address.getId());
